refact(CinematicEvent) : Cinematic event script 수정
