+++ Assets/Scripts/Event/CinematicEvent.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e6638c82e62364490b9a8095295720cd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
