+++ Assets/Scripts/Event/CinematicEvent/ImageEvents.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0b6ef15768d9247b3a23f0bc28770691
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
