+++ Assets/Scripts/Event/CinematicEvent/TextEvents.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3ea5d636101c46ac8cf0604d206bd6ea
+timeCreated: 1734871589
